% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{build_adj_mat}
\alias{build_adj_mat}
\title{Build Adjacency Matrix}
\usage{
build_adj_mat(edges, n_nodes)
}
\arguments{
\item{edges}{Edge list.}

\item{n_nodes}{Number of nodes.}
}
\value{
Adjacency matrix.
}
\description{
Build an adjacency matrix from the provided edge list.
}
\keyword{internal}
