% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet.R
\name{call_doublets}
\alias{call_doublets}
\title{call doublets by new threshold}
\usage{
call_doublets(
  scrublet_obj,
  threshold = NULL,
  python_home = Sys.which("python"),
  return_obj = TRUE
)
}
\arguments{
\item{scrublet_obj}{The result from \code{get_init_scrublet}}

\item{threshold}{Doublet score threshold for calling a transcriptome a doublet.
If \code{None}, this is set automatically.}

\item{python_home}{The python home directory where Scrublet is installed}

\item{return_obj}{If FALSE, only return a logical vector show doublet, or a python scrublet object.}
}
\value{
return a logical vector or a python scrublet object.
}
\description{
Call trancriptomes as doublets or singlets
}
