% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cat_optional}
\alias{cat_optional}
\title{Custom messaging}
\usage{
cat_optional(message, verbose)
}
\arguments{
\item{message}{message to be printed if verbose}

\item{verbose}{verbosity (bool)}
}
\description{
Custom messaging
}
