% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes by expression level and variability}
\usage{
filter_genes(
  E,
  base_ix = NULL,
  min_vscore_pctl = 85,
  min_counts = 3,
  min_cells = 3,
  plot = FALSE,
  sample_name = ""
)
}
\arguments{
\item{E}{A counts matrix where rows represent cells and columns represent genes.}

\item{base_ix}{Indices of cells to be used for v-score calculation (default: all cells).}

\item{min_vscore_pctl}{Minimum percentile threshold for v-scores.}

\item{min_counts}{Minimum expression counts required for a gene to be considered.}

\item{min_cells}{Minimum number of cells expressing a gene for it to be considered.}

\item{sample_name}{Character string specifying the name of the sample for the plot title.}

\item{show_vscore_plot}{Logical indicating whether to show a v-score plot.}
}
\value{
A numeric vector containing the indices of filtered genes.
}
\description{
This function filters genes based on expression level and variability using v-scores.
}
\seealso{
\code{\link{get_vscores}}
}
