% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_threshold}
\alias{find_threshold}
\title{Find the threshold between two maxima in a smoothed histogram}
\usage{
find_threshold(input, nbins = 256, max_num_iter = 100)
}
\arguments{
\item{nbins}{Number of bins for histogram}

\item{max_num_iter}{Maximum number of iterations for histogram smoothing}

\item{image}{Input image}

\item{hist}{Numeric vector representing the histogram}
}
\value{
The threshold between two maxima in the smoothed histogram
}
\description{
This function finds the threshold between two maxima in a smoothed histogram.
It iteratively smooths the histogram and then identifies the maxima,
breaking if fewer than three maxima are found.
}
\keyword{internal}
