% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_knn_graph}
\alias{get_knn_graph}
\title{Get k-Nearest-Neighbor Graph}
\usage{
get_knn_graph(
  X,
  k = 5,
  dist_metric = c("euclidean", "angular", "manhattan", "hamming"),
  approx = FALSE,
  return_edges = TRUE,
  random_seed = 0
)
}
\arguments{
\item{X}{Data matrix.}

\item{k}{Number of nearest neighbors.}

\item{dist_metric}{Distance metric for finding neighbors.}

\item{approx}{Whether to use approximate nearest neighbor search.}

\item{return_edges}{Whether to return edge list.}

\item{random_seed}{Random seed for reproducibility.}
}
\value{
If \code{return_edges} is \code{TRUE}, a list containing edge list and nearest neighbor matrix; otherwise, the nearest neighbor matrix.
}
\description{
Build k-nearest-neighbor graph and return edge list and nearest neighbor matrix.
}
\keyword{internal}
