% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_vscores}
\alias{get_vscores}
\title{Calculate v-scores and related statistics for genes in the input counts matrix}
\usage{
get_vscores(E, min_mean = 0, nBins = 50, fit_percentile = 0.1, error_wt = 1)
}
\arguments{
\item{E}{A counts matrix where rows represent cells and columns represent genes.}

\item{min_mean}{Minimum mean expression value for genes to be considered.}

\item{nBins}{Number of bins for calculating running quantiles.}

\item{fit_percentile}{Percentile used for fitting the running quantile.}

\item{error_wt}{Weight for the error function during optimization.}
}
\value{
A list containing the following components:
  \itemize{
    \item \code{v_scores}: V-scores for each gene.
    \item \code{CV_eff}: Coefficient of variation (CV) for effective counts.
    \item \code{CV_input}: Coefficient of variation (CV) for input counts.
    \item \code{gene_ix}: Indices of genes considered in the analysis.
    \item \code{mu_gene}: Mean expression values for selected genes.
    \item \code{FF_gene}: Fano factor (variance to mean ratio) for selected genes.
    \item \code{a}: Parameter 'a' obtained during optimization.
    \item \code{b}: Parameter 'b' obtained during optimization.
  }
}
\description{
This function calculates v-scores, coefficient of variation (CV), and other statistics
for genes in the input counts matrix. The v-score is an above-Poisson noise statistic
that helps assess the variability of gene expression levels.
}
\seealso{
\code{\link{runningquantile}}, \code{\link{optimize}}
}
