% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet.R
\name{plot_histogram}
\alias{plot_histogram}
\title{plot the doublet score histogram}
\usage{
plot_histogram(scrublet_obj, threshold = NULL)
}
\arguments{
\item{scrublet_obj}{The result from \code{get_init_scrublet} or \code{call_doublets}}

\item{threshold}{the threshold to select doublet. If NULL, use the threshold in \code{scrublet_obj}}
}
\value{
return a \code{ggplot2} object
}
\description{
plot the histogram of doublet scores for observed transcriptomes and simulated doublets
}
