% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{runningquantile}
\alias{runningquantile}
\title{Calculate running quantiles for given data points}
\usage{
runningquantile(x, y, p, nBins)
}
\arguments{
\item{x}{Numeric vector representing the x-axis data points.}

\item{y}{Numeric vector representing the y-axis data points.}

\item{p}{Percentile value for calculating the running quantiles.}

\item{nBins}{Number of bins for the running quantiles.}
}
\value{
A list containing the following components:
  \itemize{
    \item \code{xOut}: x-axis values for the running quantiles.
    \item \code{yOut}: y-axis values representing the running quantiles.
  }
}
\description{
This function calculates running quantiles for given x and y data points.
}
