% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubletR.R
\name{ScrubletR}
\alias{ScrubletR}
\title{Scrublet R6 Class}
\description{
The Scrublet R6 class represents a single-cell RNA-seq scrubbing tool.
}
\section{Public Fields}{

\itemize{
  \item \code{E_obs}: Observed expression matrix.
  \item \code{E_sim}: Simulated expression matrix.
  \item \code{E_obs_norm}: Normalized observed expression matrix.
  \item \code{E_sim_norm}: Normalized simulated expression matrix.
  \item \code{gene_filter}: Filter for highly variable genes.
  \item \code{embeddings}: List of embeddings.
  \item \code{total_counts_obs}: Total counts for observed cells.
  \item \code{total_counts_sim}: Total counts for simulated cells.
  \item \code{sim_doublet_ratio}: Ratio of simulated doublets to observed cells.
  \item \code{n_neighbors}: Number of neighbors for calculations.
  \item \code{expected_doublet_rate}: Expected doublet rate.
  \item \code{stdev_doublet_rate}: Standard deviation of doublet rate.
  \item \code{random_state}: Seed for reproducibility.
  \item \code{doublet_parents}: Matrix of doublet parents.
  \item \code{manifold_obs_}: Observed manifold data.
  \item \code{manifold_sim_}: Simulated manifold data.
  \item \code{doublet_scores_obs_}: Doublet scores for observed cells.
  \item \code{doublet_scores_sim_}: Doublet scores for simulated cells.
  \item \code{doublet_errors_obs_}: Doublet errors for observed cells.
  \item \code{doublet_errors_sim_}: Doublet errors for simulated cells.
  \item \code{doublet_neighbor_parents_}: Parents of doublet neighbors.
  \item \code{predicted_doublets}: Predicted doublets.
  \item \code{z_scores_}: Z-scores for doublet predictions.
  \item \code{threshold_}: Threshold for doublet predictions.
  \item \code{detected_doublet_rate_}: Detected doublet rate.
  \item \code{detectable_doublet_fraction_}: Detectable doublet fraction.
  \item \code{overall_doublet_rate_}: Overall doublet rate.
}
}

\section{Public Methods}{

\itemize{
  \item \code{get_dims}: Display dimensions of expression matrices.
  \item \code{scrub_doublets}: Perform scrubbing to identify and remove doublets.
  \item \code{simulate_doublets}: Simulate doublets based on observed data.
  \item \code{set_manifold}: Set manifold data.
  \item \code{calculate_doublet_scores}: Calculate doublet scores using nearest neighbors.
  \item \code{call_doublets}: Identify doublets based on scores and threshold.
  \item \code{nearest_neighbor_classifier}: Nearest neighbor classification for doublet scores.
  \item \code{plot_histogram}: Plot histogram of doublet scores.
  \item \code{set_embedding}: Set embedding data.
  \item \code{plot_embedding}: Plot embeddings.
  \item \code{pipeline_normalize}: Normalize total counts.
  \item \code{pipeline_get_gene_filter}: Identify highly variable genes.
  \item \code{pipeline_apply_gene_filter}: Apply gene filter to expression matrices.
  \item \code{pipeline_mean_center}: Mean center expression matrix.
  \item \code{pipeline_normalize_variance}: Variance normalization of expression matrices.
  \item \code{pipeline_zscore}: Z-score normalization of expression matrices.
  \item \code{pipeline_log_transform}: Log transform expression matrices.
  \item \code{pipeline_truncated_svd}: Truncated Singular Value Decomposition.
  \item \code{pipeline_pca}: Principal Component Analysis.
}
}

\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scrublet-get_dims}{\code{ScrubletR$get_dims()}}
\item \href{#method-Scrublet-new}{\code{ScrubletR$new()}}
\item \href{#method-Scrublet-scrub_doublets}{\code{ScrubletR$scrub_doublets()}}
\item \href{#method-Scrublet-simulate_doublets}{\code{ScrubletR$simulate_doublets()}}
\item \href{#method-Scrublet-set_manifold}{\code{ScrubletR$set_manifold()}}
\item \href{#method-Scrublet-calculate_doublet_scores}{\code{ScrubletR$calculate_doublet_scores()}}
\item \href{#method-Scrublet-call_doublets}{\code{ScrubletR$call_doublets()}}
\item \href{#method-Scrublet-nearest_neighbor_classifier}{\code{ScrubletR$nearest_neighbor_classifier()}}
\item \href{#method-Scrublet-plot_histogram}{\code{ScrubletR$plot_histogram()}}
\item \href{#method-Scrublet-set_embedding}{\code{ScrubletR$set_embedding()}}
\item \href{#method-Scrublet-plot_embedding}{\code{ScrubletR$plot_embedding()}}
\item \href{#method-Scrublet-pipeline_normalize}{\code{ScrubletR$pipeline_normalize()}}
\item \href{#method-Scrublet-pipeline_get_gene_filter}{\code{ScrubletR$pipeline_get_gene_filter()}}
\item \href{#method-Scrublet-pipeline_apply_gene_filter}{\code{ScrubletR$pipeline_apply_gene_filter()}}
\item \href{#method-Scrublet-pipeline_mean_center}{\code{ScrubletR$pipeline_mean_center()}}
\item \href{#method-Scrublet-pipeline_normalize_variance}{\code{ScrubletR$pipeline_normalize_variance()}}
\item \href{#method-Scrublet-pipeline_zscore}{\code{ScrubletR$pipeline_zscore()}}
\item \href{#method-Scrublet-pipeline_log_transform}{\code{ScrubletR$pipeline_log_transform()}}
\item \href{#method-Scrublet-pipeline_truncated_svd}{\code{ScrubletR$pipeline_truncated_svd()}}
\item \href{#method-Scrublet-pipeline_pca}{\code{ScrubletR$pipeline_pca()}}
\item \href{#method-Scrublet-clone}{\code{ScrubletR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$get_dims()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$new(
  counts_matrix,
  total_counts = NULL,
  sim_doublet_ratio = 2,
  n_neighbors = NULL,
  expected_doublet_rate = 0.1,
  stdev_doublet_rate = 0.02,
  random_state = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-scrub_doublets"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-scrub_doublets}{}}}
\subsection{Method \code{scrub_doublets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$scrub_doublets(
  synthetic_doublet_umi_subsampling = 1,
  use_approx_neighbors = TRUE,
  distance_metric = "euclidean",
  get_doublet_neighbor_parents = FALSE,
  min_counts = 3,
  min_cells = 3,
  min_gene_variability_pctl = 85,
  log_transform = FALSE,
  mean_center = T,
  normalize_variance = T,
  n_prin_comps = 30,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-simulate_doublets"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-simulate_doublets}{}}}
\subsection{Method \code{simulate_doublets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$simulate_doublets(
  sim_doublet_ratio = NULL,
  synthetic_doublet_umi_subsampling = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-set_manifold"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-set_manifold}{}}}
\subsection{Method \code{set_manifold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$set_manifold(manifold_obs, manifold_sim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-calculate_doublet_scores"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-calculate_doublet_scores}{}}}
\subsection{Method \code{calculate_doublet_scores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$calculate_doublet_scores(
  use_approx_neighbors = TRUE,
  distance_metric = "euclidean",
  get_doublet_neighbor_parents = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-call_doublets"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-call_doublets}{}}}
\subsection{Method \code{call_doublets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$call_doublets(threshold = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-nearest_neighbor_classifier"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-nearest_neighbor_classifier}{}}}
\subsection{Method \code{nearest_neighbor_classifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$nearest_neighbor_classifier(
  k = 40,
  use_approx_nn = TRUE,
  distance_metric = "euclidean",
  exp_doub_rate = 0.1,
  stdev_doub_rate = 0.03,
  get_neighbor_parents = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-plot_histogram"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-plot_histogram}{}}}
\subsection{Method \code{plot_histogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$plot_histogram(
  scale_hist_obs = "log",
  scale_hist_sim = "linear",
  fig_size = c(8, 3)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-set_embedding"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-set_embedding}{}}}
\subsection{Method \code{set_embedding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$set_embedding(embedding_name, coordinates)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-plot_embedding"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-plot_embedding}{}}}
\subsection{Method \code{plot_embedding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$plot_embedding(
  embedding_name,
  score = "raw",
  marker_size = 5,
  order_points = FALSE,
  fig_size = c(8, 4),
  color_map = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_normalize"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_normalize}{}}}
\subsection{Method \code{pipeline_normalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_normalize(postnorm_total = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_get_gene_filter"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_get_gene_filter}{}}}
\subsection{Method \code{pipeline_get_gene_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_get_gene_filter(
  min_counts = 3,
  min_cells = 3,
  min_gene_variability_pctl = 85,
  plot = T
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_apply_gene_filter"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_apply_gene_filter}{}}}
\subsection{Method \code{pipeline_apply_gene_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_apply_gene_filter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_mean_center"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_mean_center}{}}}
\subsection{Method \code{pipeline_mean_center()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_mean_center()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_normalize_variance"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_normalize_variance}{}}}
\subsection{Method \code{pipeline_normalize_variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_normalize_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_zscore"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_zscore}{}}}
\subsection{Method \code{pipeline_zscore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_zscore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_log_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_log_transform}{}}}
\subsection{Method \code{pipeline_log_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_log_transform(pseudocount = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_truncated_svd"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_truncated_svd}{}}}
\subsection{Method \code{pipeline_truncated_svd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_truncated_svd(
  n_prin_comps = 30,
  random_state = 0,
  algorithm = "arpack"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-pipeline_pca"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-pipeline_pca}{}}}
\subsection{Method \code{pipeline_pca()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$pipeline_pca(
  n_prin_comps = 50,
  random_state = 0,
  svd_solver = "arpack"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scrublet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scrublet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScrubletR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
