% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sparse_multiply}
\alias{sparse_multiply}
\title{Multiply each row of a sparse matrix by a scalar}
\usage{
sparse_multiply(E, a)
}
\arguments{
\item{E}{A sparse matrix where rows represent observations and columns represent features.}

\item{a}{A scalar value to multiply each row.}
}
\value{
A sparse matrix with each row multiplied by the scalar value.
}
\description{
This function multiplies each row of a sparse matrix by a scalar value.
}
\keyword{internal}
