% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sparse_zscore}
\alias{sparse_zscore}
\title{Z-score normalize each column of a sparse matrix}
\usage{
sparse_zscore(E, gene_mean = NULL, gene_stdev = NULL)
}
\arguments{
\item{E}{A sparse matrix where rows represent observations and columns represent features.}

\item{gene_mean}{A vector of mean values for each feature (default is calculated from the matrix).}

\item{gene_stdev}{A vector of standard deviation values for each feature (default is calculated from the matrix).}
}
\value{
A z-score normalized sparse matrix.
}
\description{
This function z-score normalizes each column of a sparse matrix.
}
\keyword{internal}
