% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subsample_counts}
\alias{subsample_counts}
\title{Subsample counts in a sparse matrix}
\usage{
subsample_counts(E, rate, original_totals, random_seed = 0)
}
\arguments{
\item{E}{A sparse matrix where rows represent observations and columns represent features.}

\item{rate}{Subsampling rate for counts (values between 0 and 1).}

\item{original_totals}{A vector of original total counts for each observation.}

\item{random_seed}{An integer specifying the random seed for reproducibility.}
}
\value{
A list containing the subsampled sparse matrix (E) and the final downsampling totals.
}
\description{
This function subsamples counts in a sparse matrix based on a given rate.
}
\keyword{internal}
