% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{threshold_minimum}
\alias{threshold_minimum}
\title{Return threshold value based on minimum method.}
\usage{
threshold_minimum(image = NULL, nbins = 256, max_num_iter = 10000, hist = NULL)
}
\arguments{
\item{image}{(M, N[, ...]) numeric vector, optional
Grayscale input image.}

\item{nbins}{int, optional
Number of bins used to calculate histogram. This value is ignored for
integer arrays.}

\item{max_num_iter}{int, optional
Maximum number of iterations to smooth the histogram.}

\item{hist}{array, or 2-tuple of arrays, optional
Histogram to determine the threshold from and a corresponding array
of bin center intensities. Alternatively, only the histogram can be
passed.}
}
\value{
threshold float
    Upper threshold value. All pixels with an intensity higher than
    this value are assumed to be foreground.
}
\description{
The histogram of the input \code{image} is computed if not provided and
smoothed until there are only two maxima. Then the minimum in between is
the threshold value.
}
\details{
Either image or hist must be provided. In case hist is given, the actual
histogram of the image is ignored.
}
\references{
C. A. Glasbey, "An analysis of histogram-based thresholding
    algorithms," CVGIP: Graphical Models and Image Processing,
    vol. 55, pp. 532-537, 1993.
    Prewitt, JMS & Mendelsohn, ML (1966), "The analysis of cell
    images", Annals of the New York Academy of Sciences 128: 1035-1053
    :DOI:`10.1111/j.1749-6632.1965.tb11715.x`
}
