% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tot_counts_norm}
\alias{tot_counts_norm}
\title{Total Counts Normalization}
\usage{
tot_counts_norm(
  E,
  total_counts = NULL,
  exclude_dominant_frac = 1,
  included = integer(0),
  target_total = NULL
)
}
\arguments{
\item{E}{Counts matrix to be normalized}

\item{total_counts}{Vector of total counts per cell (if NULL, computed from E)}

\item{exclude_dominant_frac}{Exclude overly abundant genes fraction (default: 1)}

\item{included}{Indices of genes to include in normalization (if empty, all genes are considered)}

\item{target_total}{Target total for normalization (if NULL, the mean of total counts is used)}
}
\value{
List containing normalized counts matrix (Enorm), average total counts, and vector of included genes
}
\description{
Total Counts Normalization
}
