% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validate_image_histogram}
\alias{validate_image_histogram}
\title{Helper function to validate image histogram.}
\usage{
validate_image_histogram(image, hist, nbins)
}
\arguments{
\item{image}{numeric vector
Grayscale input image.}

\item{hist}{array, or 2-tuple of arrays
Histogram to determine the threshold from and a corresponding array
of bin center intensities.}

\item{nbins}{int
Number of bins used to calculate histogram.}
}
\value{
list
    A list containing counts and bin_centers.
}
\description{
Helper function to validate image histogram.
}
\keyword{internal}
